hello/requirements.txt
Django==3.1.1
django-environ==0.4.5
psycopg2==2.8.6
WeasyPrint
djangorestframework